Move tests to reflect namespace change
